Fixed broken tests and added aaa cycle
Some tests that were dependent on a call to datetime.date.today were
failing because the expectations were set to the day the tests were run.
We are now mocking today's date instead. Additionally, added comments
to all the tests to illustrate the arrange-act-assert cycle.

All of this changes were made in order to fix and improve the clarity of
tests.