Added first js tests and automated way of running all the tests
Added the first passing tests for the client code in javascript. I decided
to use jest in order to write javascript tests. In addition, added more handy
commands to manage.py in order to make running of tests easier. All of this
changes were made in order to improve testing in the front end and to move
towards continuous deployment.