Changed frontend library and wrote first test
Gave up with aurelia and moved to Vue.js. Could not be happier. Wrote a test
for the first component and built it making sure it communicates with the
backend. This change was made because setting up aurelia for testing was not
working so I decided to switch to a different framework. Vue has proven
a lot more lightweight and easy to learn.