Home page redirects using auth mixin
Abstracts login for checking logged in users into a global
mixin that can be used by any component. Uses this in the home
component and adds test to make sure things are working
properly. Simplified login service interface and refactored
tests to not use done. Fixed logic to logout client when the
token has expired.